+++ tests/test_cases/missing_newline.py
@@ -1,4 +1,5 @@
 # appnexus cryptography edited google pep8 smarkets
 import ast
+# This comment should not prevent the I201 below, it is not a newline.
 import X # I201
 import flake8_import_order # I201